model/requirements.txt
flask
matplotlib
pillow
torch
